[lib] Support encrypted media types in updateMultimediaMessageMediaAction
Summary:
This diff modifies the reducer by adding two cases to the `updateMultimediaMessageMediaAction`: replacing `photo` with `encrypted_photo` and `video` with `encrypted_video`. The core part is to get rid of `uri` and `localMediaSelection` and then insert `holder` and `encryptionKey` properties.

This is needed to dynamically replace local message with the encrypted message from server once upload succeeds

Depends on D7168

Test Plan: Used redux-devtools and ensured that properties are replaced as expected/

Reviewers: ashoat, atul, marcin

Subscribers: tomek

Differential Revision: https://phab.comm.dev/D7235