[native] Add component to display encrypted images
Summary:
This diff adds a component that is able to decrypt and display encrypted images. It is similar to the `<RemoteImage>` component but I rewrote it to hooks.
The decryption is performed on every mount so this is very inefficient, caching will be added later.

Depends on D7224, D7232

Test Plan: The component itself is tested together with subsequent diffs where it is used.

Reviewers: atul, marcin, ashoat, inka

Reviewed By: ashoat

Subscribers: ginsu, tomek

Differential Revision: https://phab.comm.dev/D7225