[native] Encrypt files during media mission
Summary:
This is final diff that allows sending encrypted messages on native. The `shouldEncryptMedia` boolean has been introduced as a feature flag. Now it's hardcoded to false, but we want to enable it only for Comm community ([[ https://linear.app/comm/issue/ENG-3526/hide-media-encryption-behind-a-boolean-flag | ENG-3526 ]]) - this will be done in another diff.

The whole function becomes kind of spaghetti - a refactor follow-up wouldn't be a bad idea

Depends on D7234, D7235

Test Plan:
Sending multimedia works as before when the flag is false.
Set it to true, then send both photos and videos. Ensure both work as expected

Reviewers: atul, marcin, ashoat

Reviewed By: atul, ashoat

Subscribers: tomek

Differential Revision: https://phab.comm.dev/D7236