[native] Set SQLite DB password to `DEBUG` for `DEBUG` builds
Summary:
Interacting with the SQLite DB can be a huge pain since the path and password change on every clean install.

In the past, if you wanted to open a SQLite DB in TablePlus/DataGrip/etc, you needed to add some logging that printed `SQLiteQueryExecutor::encryptionKey` in addition to the `DB PATH`. This makes it easier and harcodes the password to `DEBUG` for `DEBUG` builds.

Test Plan: Tested iOS debug/release build and observed that `DEBUG` unlocked the SQLite DB under debug, but NOT under release.

Reviewers: ashoat, marcin, bartek, rohan

Reviewed By: ashoat

Subscribers: tomek

Differential Revision: https://phab.comm.dev/D7243