[desktop] Include generated native windows notifs
Summary:
This module was generated with [NodeRT](https://github.com/NodeRT/NodeRT) which generates native node modules from Windows API definitions. It doesn't build yet, changes required for it to  work will be included in the later diff. The module is ignored for now in the monorepo and isn't built when running `yarn` commands.

It was generated using the UI version of the NodeRT:
- winMD file was `Microsoft.WindowsAppSDK.1.2.221109.1/lib/uap10.0/Microsoft.Windows.PushNotifications.winmd`, where `Microsoft.WindowsAppSdk.1.2.21109.1` is a nuget package
- in namespaces I selected `microsoft.windows.notification`
- I unselected generate "Generate TypeScript" and "Build module"

Test Plan:
Check if it doesn't break `yarn install`

Reviewers: ashoat, tomek, marcin

Subscribers: