[desktop] Native module for windows notifications
Summary:
Changes to the generated versions:

- When running `yarn install` the new module will install the required dependencies from `nuget`
  - `packages.config` defines the nuget dependencies
  - This module is marked as optional so on macOS, `nuget install` script fails and the module is correctly ignored
- `bindings.gyp` configation file has been changed because the the automatically generated one didn't work in our unpackaged app
- There were a few changes needed in the generated c++ code itself

Also the module was added as an optional dependency of `desktop`. It's was not added to the workspace in root `package.json` because it requires `nuget` to build correctly and you can't have "optional" workspaces, they must all build correctly.

Depends on D7250

Test Plan:
On both Windows and macOS:
- `yarn cleaninstall`
- in `desktop`: `yarn package`
- try running the resulting applications and check if they works

Additionaly, for the Windows app, check with code that will be introduced in the later diff if the WNS notification channel is created.

Reviewers: ashoat, tomek, marcin

Subscribers: atul

Differential Revision: https://phab.comm.dev/D7218