[desktop] Modify the generated native windows notifs module
Summary:
Changes to the generated versions:

- When running yarn install the new module will install the required dependencies from nuget
  - packages.config defines the nuget dependencies
  - This module is marked as optional so on macOS, nuget install script fails and the module is correctly ignored
- bindings.gyp configation file has been changed because the the automatically generated one didn't work in our unpackaged app
- There were a few changes needed in the generated c++ code itself

Also the module was added as an optional dependency of desktop. It's was not added to the workspace in root package.json because it requires nuget to build correctly and you can't have "optional" workspaces, they must all build correctly.

Test Plan:
On both Windows and macOS (to check if it still builds correctly):
- `yarn cleaninstall`
- in `desktop`: `yarn package`
- try running the resulting applications and check if they work

Additionally, for the Windows app, check using code that will be introduced in the later diff if the WNS notification channel is correctly created.

Reviewers: ashoat, tomek, marcin

Reviewed By: ashoat

Subscribers: atul

Differential Revision: https://phab.comm.dev/D7218