[desktop] Embed windows runtime installer
Summary:
ummary:
We need to run the Windows runtime installer in case the user doesn't have the correct windows api dependencies (in the "packaged" apps this is specified in the app manifest but we distribute just an `.exe` so we can't use that). This diff downloads an installer from microsoft and embeds it inside our app. Then during installation the `squirrel` framework run our app in case we need to do some initialization (only on windows). Currently we just setup some shortcuts but now we will also run the runtime installer.
I've included `gaxios` dependency to make the https requests easier, we already use it in some scripts.

Test Plan:
- `yarn make` to create an installer
- Run `get-appxpackage *appruntime*` to check that the dependiecies aren't already installed (this command was taken from this [Microsoft tutorial](https://learn.microsoft.com/en-us/windows/apps/windows-app-sdk/remove-windows-app-sdk-versions))
- Install the app and check that it works
- Run `get-appxpackage` command again and check if the components were installed

Reviewers: tomek, kamil, ashoat

Reviewed By: ashoat

Subscribers: ashoat, atul

Differential Revision: https://phab.comm.dev/D7253