[native] Implement filesystem media cache
Summary:

This diff implements native methods to manage media cache stored in filesystem: the `tmp_dir/media-cache` directory.

Depends on D7251

Test Plan:

Tested together with parent diff - saved some files to cache and then retrieved it. It is possible to print the cache path and open it with Finder to inspect what files are inside.

Reviewers: atul, marcin, kamil

Subscribers: