[identity] config for LocalStack
Summary:
we need to introduce some more config for our DDB client to be able to connect to the LocalStack endpoint.

If the LOCALSTACK_ENDPOINT env var is set, the DDB client will connect to that endpoint instead of the AWS prod endpoint.

Test Plan:
called GenerateNonce with the env var set, used AWS cli to confirm that the nonce was added to my LocalStack DDB identity-nonces table
called with the env var unset, and the nonce was added to my actual DDB test table

Reviewers: jon, bartek

Reviewed By: bartek

Subscribers: ashoat, tomek, atul

Differential Revision: https://phab.comm.dev/D7258