[keyserver/lib] Optimistically create the toggle pin message
Summary:
We want to optimistically create the toggle pin message for the client pinning / unpinning the message.

Linear: https://linear.app/comm/issue/ENG-3542/optimistically-create-the-toggle-pin-message

Depends on D7220

Test Plan:
Log onto an admin user and pin a message. Confirm that the message is created locally (note: the message is pinned via opening the media gallery just to test)

{F452313}

Reviewers: atul, ginsu, tomek

Reviewed By: atul

Subscribers: ashoat

Differential Revision: https://phab.comm.dev/D7237