[desktop] Handle windows notifications
Summary:
This diff adds handling of the Windows (WNS) notifications to the desktop client. It uses the native windows node module, built in the previous diffs.

Depends on D7286, D7253

Test Plan: Run the app, connect to the keyserver and check if it correctly gets the notifications. Check if it redirects to the app after clicking on the notification.

Reviewers: kamil, tomek

Subscribers: ashoat, atul

Differential Revision: https://phab.comm.dev/D7287