[web] Add function to decrypt media
Summary: Web counterpart of D7224. It slightly differs from the native counterpart. A key difference is that we save it as a blob and generate object URL for that.

Test Plan: Called this function giving url to encrypted media. Set returned url as an image source. Confirmed the image is displayed

Reviewers: atul, marcin, kamil

Subscribers: ashoat, tomek

Differential Revision: https://phab.comm.dev/D7268