[lib/web] Add pin/unpin icon in the tooltip
Summary: We need to display an option in the tooltip to allow users to pin / unpin messages on web.

Linear: https://linear.app/comm/issue/ENG-3445/add-a-pinunpin-icon-in-the-tooltip

These are not the final designs for the pin icons in the tooltip, I've made a task to track adding those icons into the codebase: https://linear.app/comm/issue/ENG-3589/create-the-pin-icons-in-the-codebase

Test Plan: Hover over a pinned message and check to see the unpin icon, and vice versa for unpinned messages.

Reviewers: atul, ginsu, tomek

Subscribers: