[web] Add component to display encrypted media
Summary:

This diff adds a `<EncryptedMultimedia>` component that is able to decrypt and display encrypted image or video.
A loading indicator is displayed during decryption and error mark on error.
This required slight CSS adjustments, but they don't affect existing multimedia

Depends on D7268

Test Plan:

Existing non-encrypted media is not affected. New component is tested along with subsequent diffs.

Reviewers: atul, marcin, inka, kuba

Subscribers: