[web] Modify MultimediaModal to support encrypted media
Summary:
This diff modifies the `<MultimediaModal>` component to be able to display encrypted media.
The `type` and `uri` props are replaced by a single object representing an encrypted or non-encrypted media.

Depends on D7279

Test Plan: Clicking on both encrypted and non-encrypted photo (the latter one possible together with the child diff) opens a modal and displays media.

Reviewers: atul, marcin, inka, ginsu

Reviewed By: atul

Subscribers: ashoat, tomek

Differential Revision: https://phab.comm.dev/D7280