[web] Support encrypted media in <Multimedia>
Summary:
This diff modifies the `<Multimedia>` component to be able to display encrypted media.
The `type` and `uri` props are replaced by a single object representing an encrypted or non-encrypted media.

Depends on D7280

Test Plan: Both encrypted and decrypted media are displayed in the chat. Clicking on them opens a modal.

Reviewers: atul, marcin, inka, ginsu

Subscribers: ashoat, tomek

Differential Revision: https://phab.comm.dev/D7281