[keyserver] Add node-fetch package
Summary:
Adds `node-fetch` package for making requests to the WNS. It's a `2.x.x` version and not `3.x.x` because the newest version fails on `jest` tests because it's ESM-only. A lot of people are facing this issue: [SO post](https://stackoverflow.com/questions/72641681/how-to-use-node-fetch-with-jest-and-typescript), [discussion on gh](https://github.com/node-fetch/node-fetch/discussions/1503). I've tried a few workarounds but I couldn't get it to work.

- `2.x.x` is still getting security updated and bug fixes (the version I added here is from january 2023)
- there are no `flow-typed` types for `3.x.x` versions
- fixing this issue might be connected to [ENG-3414](https://linear.app/comm/issue/ENG-3414)
- we are planning to move to `fetch` anyway after updating the node version [D7253](https://phab.comm.dev/D7253#inline-47706)

Test Plan:
`yarn cleaninstall`, check if CI passes

Reviewers: ashoat, tomek

Subscribers: