Extend CommIOSNotification module with method to send an event to JS to trigger message infos persistence
Summary: This differential adds method to CommIOSNotifications module that can send event to JS to persist in redux message infos that wasreceived by Notification Service Extension while the app was in the background.

Test Plan: Build the app. Call this method in `didReceiveNotification` callback in AppDelegate with dummy message infos. Subscribe in JS to log message infos to console.

Reviewers: tomek, bartek

Subscribers: