Send Darwin Notification from NSE to AppDelegate so that it consumes flat file storage immediately and sends event to JS
Summary: This differential makes NSE send Darwin Notification to AppDelegate when it receives the notif. AppDelegate reacts by consuming the flat file and sending event to JS to persist message infos.

Test Plan: Build the app. Background it. Send notification. Kill the keyserver (debug)/turn off network connection (release). Foreground the app. App is dosconnected from the keyserver but the message is visible in the app.

Reviewers: tomek, bartek

Reviewed By: tomek

Subscribers: ashoat, atul

Differential Revision: https://phab.comm.dev/D7285