[native] Add `avatar` column to clientDB `threads` table
Summary:
We are adding an `avatar` column to the `threads` table in the `clientDB` and including an `avatar` field in the `Thread` struct. At the moment, the value of this field will always be `NULL` until we make some necessary changes. These changes involve modifying:

1. The translation functions that convert between `ClientDBThreadInfo` and `ThreadInfo`.
2. The C++ code responsible for constructing JSI `ClientDBThreadInfo` objects from the `threads` table rows retrieved from the database on the "JS side."

Once these changes have been made, we will be able to update the `avatar` field in the `threads` table and properly persist data in the new `avatar` column.

Test Plan:
1. Verify that the migration was successful:

Before:
{F455939}

After:
{F455940}

{F455941}

2. Ensure that freshly created `clientDB` includes `avatar` column in the `threads` table:

Reviewers: ashoat, ginsu, kamil

Reviewed By: ashoat, kamil

Subscribers: tomek

Differential Revision: https://phab.comm.dev/D7257