[nix] Set db variables in our local environments
Summary:
issue: https://linear.app/comm/issue/ENG-3349/test-setting-innodb-ft-min-token-size
In D7266 we set `innodb-ft-min-token-size` and `innodb-ft-enable-stopword` variables in the db run in Docker. We also need a way for these variable to be set in our local environments. This diff addresses the Nix
MariaDB instances. The legacy instances will have to be updated manually, and a proper instruction will be provided in the linear task, as well as on the Comm Dev Team channel.
I would be grateful if someone could check if the messages printed by `db_usage` and `usage` functions are reasonable.

Test Plan:
Running a nix db instance, I ran `comm-dev db restart` and checked that the variables were set to 1 and 0, respectively.

Reviewers:
jon, ashoat
Subscribers: