[native] Display edited message content in the pending threads
Summary:
~~Previously: [fix] Fix 'Maximum update depth exceeded' on native when opening tooltip~~
Can't open the tooltip menu, because of maximum update depth exceeded warning.
Issue: https://linear.app/comm/issue/ENG-3584/cant-open-tooltips-on-native

Changed the way in which the updated content in the pending thread was displayed. Now, a new edit message is added to the pending thread, instead of modifying the existing one.

Test Plan: Run the app on iOS, open the tooltip menu, it should work. Checked on web if everything still works. Checked if height measurement doesn't display warnings in the console.

Reviewers: michal, inka, ginsu

Reviewed By: inka

Subscribers: ashoat, tomek, atul

Differential Revision: https://phab.comm.dev/D7301