[web] gate avatars rendering
Summary:
Added a hard variable called `shouldRenderAvatars` to gate the avatars rendering work for web. Gating this work will allow me to land all the diffs in this stack without launching avatars. I know this is a big diff; however, when we eventually have avatars in stable condition, I think it will be nice to get rid of this code, so having one diff that shows all the changes that need to be "reverted" is why I decided to make this diff bigger than normal.

Depends on D7276

Test Plan:
Tested all areas that have avatars and made sure that the avatars didn't render when the gate was closed:

When the gate is closed:
{F459895}

When the gate is open:
{F459896}

Reviewers: ashoat, atul

Reviewed By: atul

Subscribers: ashoat, tomek, atul

Differential Revision: https://phab.comm.dev/D7277