[native] introduce ColorRows component
Summary: For the emoji avatar creation flow we want the same color rows logic that are found in `ColorSelector`. I factored out this logic into its own component so that we can better resuse it in both places

Test Plan:
There were no regressions, and `ColorSelector` looks and behaves as it did previously.

{F462922}

Reviewers: ashoat, atul

Reviewed By: ashoat

Subscribers: ashoat, tomek, atul

Differential Revision: https://phab.comm.dev/D7302