[keyserver] Create an endpoint that can be used to validate an invite link
Summary:
This endpoint checks if the link is valid and returns community details if that's the case. Currently, link expiration is not checked. Avatars will be returned in the future, after their API stabilises.

Depends on D7519

Test Plan: Call the endpoint from the client and check if it returns correct value for links that are valid, invalid and that points to a community to which a user already belongs.

Reviewers: bartek, kamil, inka