[keyserver] Move deep link handling logic to a separate function
Summary:
This function grows and now it makes sense to extract it.
Depends on D7271

Test Plan: Checked on a physical iPhone that clicking this link can open the App Store. It's just one branch of the logic, but this diff is only about moving the code, so it should suffice.

Reviewers: bartek, kamil

Subscribers: ashoat, atul

Differential Revision: https://phab.comm.dev/D7311