[keyserver] Create a page with invite link
Summary:
When a user doesn't have the app installed on iOS, clicking the invite link will open this page. User should use a link to App Store to install the app and click a link again to open the invitation.
The design was based on https://linear.app/comm/issue/DES-49#comment-71838f8d but the values were guessed as only images were shared. They will be updated in next iterations.
Community details section is just a placeholder as it requires some backend work before we can match a link with a community.
The font will be updated in the next diff.

{F463770}

Depends on D7311

Test Plan: Open the page on simulator and on a physical iPhone and check if links work correctly.

Reviewers: bartek, kamil

Subscribers: ashoat, atul

Differential Revision: https://phab.comm.dev/D7312