[keyserver] Create a page with invite link
Summary:
When a user doesn't have the app installed on iOS, clicking the invite link will open this page. User should use a link to App Store to install the app and click a link again to open the invitation.

The design doesn't use correct Comm title font, images and link icon - these will be added in next diffs.

{F467738}

Depends on D7311

Test Plan: Open the page on simulator and on a physical iPhone and check if links work correctly.

Reviewers: bartek, kamil

Reviewed By: kamil

Subscribers: ashoat, atul

Differential Revision: https://phab.comm.dev/D7312