[keyserver] Use the same font as on web
Summary:
Initially I thought I have to do something similar to what we do on web - call `getFontsURL` and include it as a stylesheet. But after reading some code I realized that that font file contains only `IBM Plex Sans` font family which isn't a part of our font stack.

Depends on D7312

Test Plan: Open the page on simulator and on a physical iPhone and check if the new font is displayed.

Reviewers: bartek, kamil