[native] Added edit button to the Tooltip
Summary:
Added Edit button to the tooltip, which selects a message for
editing. It is shown only for user-created messages.

Test Plan:
Checked if the edit button displays correctly in the tooltip.
Checked if the buttons are correctly handled if there are many of them.
Checked if the state is correctly changed in devtools.

TODO: // screenshots

Reviewers: michal, inka, ginsu

Subscribers: ashoat, tomek, atul

Differential Revision: https://phab.comm.dev/D7343