[native] Added edit button to the Tooltip
Summary: Added Edit button to the tooltip, which selects a message for editing. It is shown only for messages in which the author is the viewer. The edit button is hidden behind the `EDIT_BUTTON_DISPLAY` feature flag.

Test Plan:
Checked if the edit button displays correctly in the tooltip.
Checked if the buttons are correctly handled and if there are many of them.
Checked if the state is correctly changed in dev tools.
Checked if the button is only shown for users from `staff.js`.

{F468040}

Reviewers: michal, inka, ginsu

Reviewed By: inka

Subscribers: ashoat, tomek, atul

Differential Revision: https://phab.comm.dev/D7343