[native] Change 'addReply' function to 'editInputMessage'
Summary:
For now, prepending text to add a reply was enough. For editing messages, I will need also an option to replace text in the chat input bar. I renamed the function and added a new parameter (mode), to handle the new situation. The usage of new mode will be added in incomming
diffs.

Test Plan: Run the app on iOS, and check if the reply mode still works. In the next diff checked if the new mode ('replace') also works when entering edit mode.

Reviewers: michal, inka, ginsu

Reviewed By: inka

Subscribers: ashoat, tomek, atul

Differential Revision: https://phab.comm.dev/D7345