[native] Copy the edited message content to ChatInputBar when entering edit mode
Summary: We want to set the content of the edited message in the ChatInputBar when entering edit mode. We need a callback, to always have the latest state.

Test Plan: Run the app on iOS. Checked if when entering the edit mode the content of the message is copied to the InputBar.

Reviewers: michal, inka, ginsu

Subscribers: