[native] Preserve the draft for ChatInputBar after leaving edit mode
Summary: We want to preserve user input after exiting edit mode. If we are in edit mode we don't override the draft. After exiting it we want to restore the previous content. We need a callback, to always have the latest state.

Test Plan: Run the app on iOS. Typed a message in the input bar. Entered edit mode. Changed the content of the message in the input bar. Pressed cross to exit edit mode. The draft was restored correctly.

Reviewers: michal, inka, ginsu

Reviewed By: inka

Subscribers: ashoat, tomek, atul

Differential Revision: https://phab.comm.dev/D7347