[keyserver] Add an arrow next to a link
Summary:
Add a svg arrow icon next to Comm landing link. The svg was exported from figma and added directly to this file - this was the quickest solution.

Depends on D7348

Test Plan: Open the page on simulator and physical iPhone and check with the icon is displayed and clicking it or the link opens the landing page.

Reviewers: bartek, kamil

Subscribers: ashoat, atul

Differential Revision: https://phab.comm.dev/D7350