[keyserver] Rename importJSON to getCommConfig
Summary:
This is a simple rename diff. In addition to renaming the function, I also renamed the file it's defined in, and added a small code comment clarifying its functionality.

The old name made sense when the function only handled the JSON file scenario, but now that it also pulls from environmental variables I think the new name is better.

Test Plan: Flow

Reviewers: atul, varun, jon

Subscribers: tomek

Differential Revision: https://phab.comm.dev/D7351