[CommCoreModule] Update thread `REPLACE_OPERATION` to handle `avatar`
Summary:
We need to update construction of `REPLACE_OPERATION` to handle `avatar`.

After this diff `avatar` will be persisted in the SQLite `threads` table and will REHYDRATE Redux.

NOTE: I am NOT going to include a migration to handle `avatar`s that may already be in the Redux store because there should be no `avatar`s on prod as of this diff landing.

Test Plan:
1. Set thread avatar from `native`:

2. Observe that thread avatar is set in `serverDB`:

3. Observe that Redux state on `native` is updated with `avatar`:

4. Observe that `avatar` field is persisted in `clientDB` `threads` table:

5. Observe that `avatar` is included in REHYDRATE action:

Reviewers:
ashoat, ginsu