[native] introduce outerRingSelectedColor prop to ColorSelectorButton and ColorRows
Summary: While working on the emoji avatar creation, I noticed a situation where the outer ring of the selected color was not visible. This was because the outer ring color matched the background color of the screen area. To fix this, I created an optional prop called `outerRingSelectedColor` that devs can use to override the outer ring color for situations like this

Test Plan:
Please watch the demo videos to see the before and after

Before:
{F464364}

After (set the value of `outerRingSelectedColor` to `#404040`) :

{F464390}

Reviewers: ashoat, atul

Reviewed By: ashoat

Subscribers: ashoat, tomek, atul

Differential Revision: https://phab.comm.dev/D7321