[native] changed name of loginText to whiteText
Summary: For the emoji avatar creation screen, I need a white color value. In the `colors.js` file there is a color entry for `loginText` which is white. I noticed that it was only being used in one place, so changed the name from `loginText` to `whiteText` so that it could be better used in both places

Test Plan:
Searched for all uses of `loginText` and saw that it was used in one place. After this change, I checked the login screen and everything looked the same as it did before:

{F476192}

Reviewers: ashoat, atul

Reviewed By: ashoat

Subscribers: ashoat, tomek, atul

Differential Revision: https://phab.comm.dev/D7353