Add method to CryptoModule to generate new prekey
Summary: This differential adds method to CryptoModule to generate new prekey. It will be needed furhter in this stack to initialize session between the device and keyserver for e2e notifs with X3DH

Test Plan: Call this method in `CommCoreModule::initializeCryptoAccount` and attach debugger. Ensure that returned prekey matches the form described here:
https://github.com/CommE2E/olm/blob/main/src/account.cpp#L326

Reviewers: tomek, atul, bartek

Subscribers: