[web] Display the pinned messages for the thread when the banner is clicked
Summary: The pinned messages modal for the thread should display when the user clicks on the banner indicating the number of pinned messages for that given thread. Each message should display the inline
engagement (reactions and replies), based on the designs. This is different from the message being displayed in the toggle pin modal, as the designs there indicated that we wanted to remove the inline
engagement.

Test Plan: Pin several messages, with / without inline engagement and a mix between text and media. Confirm that the media gallery looks as expected.

Reviewers: atul, ginsu, tomek

Subscribers: