Refactor NotificationsCryptoModule. Add methods to get one time keys and generate new prekey
Summary:
This differential adds methods to NotificationsCryptoModule to generate new one time keys and new prekey respectively.
Additionally the class is refactored so that every call to in-memory `CryptoModule` instance is wrapped around file deserialization
and serialization. This is both to ensure that any change to in-memory `CryptoModule` is reflected in persistent storage and to
avoid repeated code.

Test Plan: Call each of added/modified method in `CommCoreModule.initializeCryptoAccount` with debugger attached.

Reviewers: tomek, atul, bartek

Subscribers: ashoat

Differential Revision: https://phab.comm.dev/D7383