[identity] add CompareUsers RPC to identity_client.proto
Summary: This is the last relevant RPC from the old identity.proto. We can now remove the file

The actual implementation of this API will be slightly different from the original. We will be comparing usernames instead
of user IDs

Depends on D7386

Test Plan: build the identity service successfully

Reviewers: ashoat

Subscribers: