[web] Fix chat composer when no active thread
Summary: Linear task: [ENG-3623](https://linear.app/comm/issue/ENG-3623/create-chat-button-is-broken-on-web-if-all-chats-are-unread)

Test Plan:
Before this, the chat composer wouldn't open if there was no active thread. After this, it works fine, both when pressing the compose button and when loading it directly by navigating to http://localhost:3000/comm/chat/thread/new/

Note that the only way to get to a state with no active thread is load the web app when ALL threads are unread. (This is what happens when somebody creates a new account on web)

Reviewers: inka, tomek

Reviewed By: tomek

Subscribers: atul

Differential Revision: https://phab.comm.dev/D7388