[native] Discard alert when navigating in edit mode
Summary: When we are in edit mode we don't want to lose changes by navigating. I display an alert when we are in edit mode. For now, it displays the alert always, when in edit mode.

Test Plan: Run the app on iOS, and check if in edit mode when going back to the thread list, the alert is displayed, and if works as expected.

Reviewers: michal, inka, ginsu