[native] Discard alert when navigating in edit mode
Summary: When we are in edit mode we don't want to lose changes by navigating. I display an alert when we are in edit mode and we have some changes in the edited message.

Test Plan: Run the app on iOS, and check if in edit mode with changes when going back to the thread list, the alert is displayed, and if works as expected.

Reviewers: michal, inka, ginsu

Reviewed By: michal

Subscribers: ashoat, tomek, atul

Differential Revision: https://phab.comm.dev/D7446