Introduce type for keys needed to initialize OLM session
Summary: This differential introduces types for OLM prekey and one time keys batch. Additionally it groups prekey and one time keys batch into a type that represents data needed to initialize olm
session according to X3DH

Test Plan: It is wise to run `flow` but it will not check much since those types are not used anywhere yet. Better way to test this differential is to log to console outputs of
`CryptoModule::generateNewPrekey` and `CryptoModule::getOneTimeKeys()` and ensure those outputs match presented types. Note that type for prekey is somewhat different. It was changed to make sure that
there is only one id <-> key pair inside `curve25519`.

Reviewers: atul, tomek, bartek

Subscribers: